requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
Werkzeug==2.3.7
python-dotenv==1.0.0
WTForms==3.0.1
email-validator==2.0.0
flask-cors==4.0.0

# Vision & AI
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
ultralytics==8.1.0
mediapipe==0.10.11
pyzbar==0.1.9
pytesseract==0.3.10

# Scheduler
APScheduler==3.10.4

# Email
Flask-Mail==0.9.1

# Real-time notifications (NEW)
flask-socketio==5.3.5
python-socketio==5.10.0

# Production server (NEW)
gunicorn==21.2.0
eventlet==0.36.1

# Authentication & Security (NEW)
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Validation (NEW)
pydantic==2.5.0
pydantic[email]==2.5.0

# Caching (NEW)
redis==5.0.1

# PDF Export
reportlab==4.0.7

# Telegram Bot
requests==2.31.0

# QR Code Generation
qrcode[pil]==7.4.2

# Database (PostgreSQL)
psycopg2-binary==2.9.9

# Data Science & Analytics
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2

# LLM & RAG (NEW)
torch>=2.0.0
transformers>=4.31.0
bitsandbytes>=0.41.1
accelerate>=0.21.0
chromadb>=0.4.15
sentence-transformers>=2.2.2
scipy>=1.11.0

